requirements.txt
flask==3.0.0
kafka-python==2.0.2
pandas==2.1.4
pyarrow==14.0.1
langchain-openai==0.1.19
langchain-core==0.2.38
pydantic>=2.7,<3
watchdog
pytest
numpy
python-dotenv==1.0.1
